renderer/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-pptx==1.0.2
pydantic==2.11.7
